Refactor Burst Utils
