+++ app/src/main/java/com/harrysoft/burstcoinexplorer/accounts/ui/SavedAccountsRecyclerAdapter.java
@@ -13,7 +13,7 @@ import android.widget.TextView;
 import com.harrysoft.burstcoinexplorer.R;
 import com.harrysoft.burstcoinexplorer.accounts.db.SavedAccount;
 import com.harrysoft.burstcoinexplorer.burst.entity.BurstAddress;
-import com.harrysoft.burstcoinexplorer.burst.utils.BurstUtils;
+import com.harrysoft.burstcoinexplorer.burst.util.BurstUtils;
 import com.harrysoft.burstcoinexplorer.router.ExplorerRouter;
 
 import java.util.ArrayList;
