+++ app/src/main/java/com/harrysoft/burstcoinexplorer/explore/browse/ViewTransactionDetailsActivity.java
@@ -9,8 +9,8 @@ import android.widget.Toast;
 import com.harrysoft.burstcoinexplorer.R;
 import com.harrysoft.burstcoinexplorer.burst.service.BurstBlockchainService;
 import com.harrysoft.burstcoinexplorer.burst.entity.Transaction;
-import com.harrysoft.burstcoinexplorer.burst.utils.BurstUtils;
-import com.harrysoft.burstcoinexplorer.burst.utils.TransactionTypeUtils;
+import com.harrysoft.burstcoinexplorer.burst.util.BurstUtils;
+import com.harrysoft.burstcoinexplorer.burst.util.TransactionTypeUtils;
 import com.harrysoft.burstcoinexplorer.router.ExplorerRouter;
 import com.harrysoft.burstcoinexplorer.util.TextViewUtils;
